Create localStorage module
